+++ src/main/java/com/project/coalba/global/utils/ProfileUtil.java
@@ -7,8 +7,6 @@ import com.project.coalba.domain.profile.repository.StaffProfileRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 
-import java.util.Optional;
-
 @RequiredArgsConstructor
 @Component
 public class ProfileUtil {
@@ -30,14 +28,4 @@ public class ProfileUtil {
          return staffProfileRepository.findById(staffId)
                  .orElseThrow(() -> new RuntimeException("해당 이용자가 존재하지 않습니다."));
     }
-
-    public Boss getBossByScheduleId(Long scheduleId) {
-        Boss boss = bossProfileRepository.findByScheduleId(scheduleId);
-        if(boss == null){
-            throw new RuntimeException("해당 스케줄의 사장님이 존재하지 않습니다.");
-        } else{
-            return boss;
-        }
-    }
-
 }
